@@ -18,22 +18,29 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.2
+  cupertino_icons: ^1.0.3
+  equatable: ^2.0.0
   flutter:
     sdk: flutter
   # pedantic: ^1.11.0 
   # lint: ^1.0.0
 
+  flutter_bloc: ^7.0.0
   flutter_localizations:
     sdk: flutter
-  http: ^0.13.1  
+  get_it: ^7.0.0  
+  http: ^0.13.3  
+  infinite_listview: ^1.1.0
   intl: ^0.17.0  
   path_provider: any
+  provider: ^5.0.0
+  rxdart: ^0.26.0
+  scoped_model: ^1.1.0
   shared_preferences: ^2.0.5  
   sqflite: any
   
