@@ -1,6 +1,18 @@
-import 'dart:convert';
+import 'package:equatable/equatable.dart';
 
-class Transaction {
+class Transaction extends Equatable {
+  const Transaction(
+      this.author,
+      this.department,
+      this.date,
+      this.sum,
+      this.source,
+      this.time,
+      this.type,
+      this.user,
+      this.analytics,
+      this.mathSymbol);
+
   final String author;
   final String department;
   final String date;
@@ -12,8 +24,19 @@ class Transaction {
   final String analytics;
   final String mathSymbol;
 
-  Transaction(this.author, this.department, this.date, this.sum, this.source,
-      this.time, this.type, this.user, this.analytics, this.mathSymbol);
+  @override
+  List<Object?> get props => [
+        author,
+        department,
+        date,
+        sum,
+        source,
+        time,
+        type,
+        user,
+        analytics,
+        mathSymbol
+      ];
 
   factory Transaction.fromJson(Map<String, dynamic> json) {
     return Transaction(
@@ -29,27 +52,10 @@ class Transaction {
         json['mathSymbol'] as String);
   }
 
-  String clientToJson(Transaction transaction) {
-    final transactionJson = transaction.toJson();
-    return json.encode(transactionJson);
-  }
-
   Map<String, String> toJson() => {
         'author': author,
         'department': department,
         'date': date,
         'sum': sum.toString(),
       };
-
-  // factory String.toJson(Transaction transaction) {
-  //   Map<String, Transaction>() => {
-  //         "author": author,
-  //         "department": department,
-  //         "date": date,
-  //         "sum": sum,
-  //       };
-
-  //   return "rr";
-  //   // return json.encode(d);
-  // }
 }
